+++ pure_protobuf/__init__.py
@@ -6,7 +6,3 @@ Python implementation of Protocol Buffers data types.
 
 `pure-protobuf` contributors © 2011-2019
 """
-
-from __future__ import absolute_import
-
-from pure_protobuf.protobuf import *  # noqa: F401, F403