+++ tests/test_google.py
@@ -1,0 +1,20 @@
+"""
+`pure-protobuf` contributors © 2011-2019
+"""
+
+# noinspection PyCompatibility
+from dataclasses import dataclass
+from datetime import datetime, timezone
+from typing import Optional
+
+from pure_protobuf.dataclasses_ import field, message
+
+
+def test_datetime():
+    @message
+    @dataclass
+    class Test:
+        timestamp: Optional[datetime] = field(0, default=None)
+
+    test = Test(timestamp=datetime.now(tz=timezone.utc))
+    assert Test.loads(test.dumps()) == test